feat(Forma/Geometry): add drawable_canvas geometry function and wire_canvas_drag helper
drawable_canvas produces a GeometryFn<vector<float>> rendering N samples as
a LINE_LIST polyline across NDC bounds. wire_canvas_drag registers on_drag
on a Context, mapping cursor NDC to sample index and amplitude with linear
interpolation between the previous and current index to prevent sparse
samples under fast drag. Version increment on state drives sync() and
bulk_reader routing through Bridge.
